refactor: 코드리뷰 리팩토링 반영
